cephalus_hunter/requirements.txt
flask==3.0.3
psutil==6.0.0
pywin32==306; platform_system == "Windows"
pyjwt==2.9.0
bleach==6.1.0
requests==2.32.3
flask-limiter==3.8.0